Changes from Chloe's code review
